add empty state and adjusts layout

+++ src/components/ProductList/ProductList.tsx
@@ -1,10 +1,12 @@
 import { useNavigation } from '@react-navigation/core';
-import React from 'react';
+import React, { useCallback } from 'react';
 import { FlatList } from 'react-native';
 import { useSelector, useDispatch } from 'react-redux';
 import { ProductState, actions, IProduct } from '~/store/ducks/product';
+import EmptyState from '../EmptyState/EmptyState';
 
 import ProductItem from '../ProductItem/ProductItem';
+import ProductListSkeleton from './ProductListSkeleton';
 
 const ProductList = () => {
   const productState = useSelector<{ product: ProductState }>(
@@ -29,21 +31,40 @@ const ProductList = () => {
       formData: product,
     });
   };
-  return (
-    <FlatList
-      data={productState.products}
-      keyExtractor={item => String(item.id)}
-      renderItem={({ item }) => (
-        <ProductItem
-          handleDelete={() => {
-            handleDelete(item.id);
-          }}
-          handleUpdate={() => handleUpdate(item)}
-          {...item}
+
+  const renderList = useCallback(() => {
+    if (productState.products?.length) {
+      return (
+        <FlatList
+          data={productState.products}
+          keyExtractor={item => String(item.id)}
+          renderItem={({ item }) => (
+            <ProductItem
+              handleDelete={() => {
+                handleDelete(item.id);
+              }}
+              handleUpdate={() => handleUpdate(item)}
+              {...item}
+            />
+          )}
         />
-      )}
-    />
-  );
+      );
+    }
+
+    if (productState.error) {
+      return (
+        <EmptyState
+          title="Erro"
+          message="não foi possível atender sua solicitação tente novamente."
+          retry={() => dispatch(actions.fetchProductsRequest())}
+        />
+      );
+    }
+
+    return <ProductListSkeleton />;
+  }, [productState, handleDelete, handleUpdate, dispatch]);
+
+  return renderList();
 };
 
 export default ProductList;
